requirements.txt
fastapi==0.115.12
py-clob-client==0.27.0
python-dotenv==1.1.1
requests==2.32.3
uvicorn==0.34.2
